+++ example/package.json
@@ -12,14 +12,9 @@
     "@pagopa/io-react-native-crypto": "^0.2.3",
     "@pagopa/io-react-native-integrity": "^0.1.0",
     "@pagopa/io-react-native-jwt": "^1.2.0",
-    "@react-native-community/masked-view": "^0.1.11",
-    "@react-navigation/bottom-tabs": "^6.5.20",
-    "@react-navigation/native": "^6.1.17",
-    "@react-navigation/native-stack": "^6.9.26",
     "react": "18.2.0",
     "react-native": "0.72.14",
-    "react-native-safe-area-context": "^4.9.0",
-    "react-native-screens": "^3.30.1"
+    "react-native-safe-area-context": "^4.9.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
