+++ example/src/scenarios/poc/get-credential-multiple.ts
@@ -3,7 +3,7 @@ import {
   Trust,
   createCryptoContextFor,
 } from "@pagopa/io-react-native-wallet";
-import { error, result, toResultOrReject } from "./types";
+import { error, result, toResultOrReject } from "../types";
 import getWalletInstanceAttestation from "./get-attestation";
 import { generate } from "@pagopa/io-react-native-crypto";
 import getPid from "./get-pid";
