+++ example/src/scenarios/poc/get-pid.ts
@@ -2,7 +2,7 @@ import {
   Credential,
   createCryptoContextFor,
 } from "@pagopa/io-react-native-wallet";
-import { error, result, toResultOrReject } from "./types";
+import { error, result, toResultOrReject } from "../types";
 import getWalletInstanceAttestation from "./get-attestation";
 import { generate } from "@pagopa/io-react-native-crypto";
