+++ example/src/scenarios/types.ts
